# perf: test connector perf