@Crash-- 2023-02-17
Also can you tell us: 
- On which device? 
- Android version? 
- Chrome or android system webview version 🙏