@acezard 2023-02-17
> Also can you tell us:
> 
> * On which device?
> * Android version?
> * Chrome or android system webview version 🙏

Hello, I wouldn't take those test results very seriously by the way (I mean they probably show the real proportions but not the real numbers), as the sample size is really low, and the test environment not prod-like. Still:
- Nexus_5X_API_30 / Google Play Intel Atom (x86) (AVD)
- Android 11
- System Webview (I can access it from MS Edge, so I guess it's not Chrome)