@acezard 2023-02-17
> Interesting, thanks for the first numbers.
> 
> * 15/20s to check if the user is authenticated is problematic. Can you add a timer on the `onLoad` event within the webview? To know the time taken to have the DOM Ready event and then to know how long the method to detect if the user is connected or not really takes.
> * can we "split" the `fetch` method? Can we have several numbers:
>   
>   * check goTo time (https://github.com/konnectors/template_ccc/blob/master/src/index.js#L46)
>   * check the time of the selector in the worker https://github.com/konnectors/template_ccc/blob/master/src/index.js#L47
>   * check the time of clickAndWait https://github.com/konnectors/template_ccc/blob/master/src/index.js#L48
>   * check the method of parseBills https://github.com/konnectors/template_ccc/blob/master/src/index.js#LL49C50-L49C50
>   * check the time of saveBills https://github.com/konnectors/template_ccc/blob/master/src/index.js#L52-L57
> 
> From what I see between `When creating an account with cached bundle` and `When refreshing data` I imagine that having a 22s drop between both `ensureAuthenticated` is because of the webview cache?

Adding a checklist for timings:

- [ ] onLoad event within the WebView (both worker and pilot)
- [ ] goTo (fetch method)
- [ ] selector in the worker
- [ ] clickAndWait()
- [ ] parseBills()
- [ ] saveBills()

As for your question about the difference in ensureAuthenticated between a new account and an existing account, I'm not sure what could create such a difference, maybe cache as you said. @doubleface do you have an idea, please?