@Crash-- 2023-02-20
- ajout du log sur onLoad 
- mettre debug a true sur launcherview pour voir si nos scripts interagissement dès que la page est prête visuellemetn ou si y a beaucoup de latence.